requirements.txt
asgiref==3.11.0
Django==4.2.26
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
psycopg2-binary